Merge branch 'feature/compile_with_latest_VS' into develop

+++ Builder/Builder/Window.cpp
@@ -2,6 +2,8 @@
 
 	#include "Window.h"
 
+#include <wingdi.h>
+
 
 // zewnetrzne zmienne oraz funkcje
 
@@ -23,7 +25,7 @@
 	Tworzy okno, kontekst urz¹dzenia oraz kontekst tworzenia grafiki.
 *****************************************************************************/
 
-	BOOL SetupWindow(const char *title, int width, int height, int bits, bool isFullscreen)
+	BOOL SetupWindow(const char *title, int width, int height, BYTE bits, bool isFullscreen)
 	{
 	// ustawia globalny znacznik
